Fix crash, when try to press key "Space" or "Enter" when camera hasn't been turned on yet.

+++ src/frontends/qtfrontend/qtfrontend.cpp
@@ -990,7 +990,7 @@ void QtFrontend::setToolBarState(int newState)
 
 int QtFrontend::getRecordingMode()
 {
-    return mw->getRecordingMode();
+    return static_cast<int>(mw->getRecordingMode());
 }
